模型/requirements.txt
pandas==2.0.3
numpy==1.24.3
sentence_transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
torch==2.0.1
transformers==4.30.2
accelerate==0.20.3
openpyxl==3.1.2
scikit-learn==1.3.0
python-dotenv==1.0.0
streamlit==1.28.2